Moved functions out of ServerCall.jsx and put them where used. Added a few to a new sever_utils file as well as the ServerCalls and BASE_URL variables
