+++ TelemetrySite/client/src/DataUpload/DataUpload.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 
 import "./DataUpload.css";
-import { PostDataFile, FetchProgress } from "../ServerCall/ServerCall";
+import { BuildURI, CheckData } from "../server_utils";
 import {
   Container,
   Row,
@@ -57,9 +57,73 @@ function DataUpload() {
       return;
     }
 
+    const PostDataFile = async (mf4File, dbcFile, contextData) => {
+      // Ensure CheckData() completes before proceeding
+      if (!ServerCalls) {
+        try {
+          const response = await CheckData(); // Await the result
+          if (!response) {
+            console.error("CheckData failed or returned an invalid response.");
+            return false; // Stop execution if CheckData fails
+          }
+        } catch (error) {
+          console.error("Error in CheckData:", error);
+          return false;
+        }
+      }
+    
+      const formData = new FormData();
+      formData.append("mf4File", mf4File);
+      formData.append("dbcFile", dbcFile);
+      formData.append("contextData", contextData);
+    
+      try {
+        const response = await fetch( await BuildURI(), {
+          method: "POST",
+          body: formData,
+        });
+    
+        if (!response.ok) {
+          const jsonResponse = await response.json();
+          console.error(
+            "Error occurred on server side. Error message: " + jsonResponse.error
+          );
+          return false;
+        }
+        return true;
+      } catch (error) {
+        console.error("Network or server error:", error);
+        return false;
+      }
+    };
+
     const mf4File = document.getElementById("fileUploadMF4").files[0];
     const dbcFile = document.getElementById("fileUploadDBC").files[0];
     const response = PostDataFile(mf4File, dbcFile, contextData);
+
+    /**
+     * Fetch the progress of the current upload
+     *
+     * @return {int} decimal of how much has been uploaded
+     */
+    const FetchProgress = async () => {
+      try {
+        const response = await fetch(BASE_URL + ServerCalls["data_upload"], {
+          method: "GET",
+        });
+    
+        if (!response.ok) {
+          console.error("Network response was not ok: " + response.statusText);
+        }
+    
+        const data = await response.json();
+    
+        return data;
+      } catch (error) {
+        console.error("Failed to fetch progress:", error);
+        return { error: error.message };
+      }
+    };
 
     var fileUpload = false;
     var lastProgress = -1;
